Fix standalone test errors after refactoring
