+++ functionaltests/internal/esclient/client_v7.go
@@ -93,6 +93,11 @@ func (c *Client) getDocCountV7(ctx context.Context, name string) (docCountV7, er
 		return docCountV7{}, fmt.Errorf("cannot get count for %s: %w", name, err)
 	}
 
+	// If not found, return zero count instead of error.
+	if resp.StatusCode == http.StatusNotFound {
+		return docCountV7{Count: 0}, nil
+	}
+
 	if resp.StatusCode > http.StatusOK {
 		return docCountV7{}, fmt.Errorf(
 			"count request for %s returned unexpected status code: %d",
