--- functionaltests/steps_v7.go
@@ -1,203 +1,0 @@
-// Licensed to Elasticsearch B.V. under one or more contributor
-// license agreements. See the NOTICE file distributed with
-// this work for additional information regarding copyright
-// ownership. Elasticsearch B.V. licenses this file to you under
-// the Apache License, Version 2.0 (the "License"); you may
-// not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package functionaltests
-
-import (
-	"context"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/require"
-
-	"github.com/elastic/apm-server/functionaltests/internal/asserts"
-	"github.com/elastic/apm-server/functionaltests/internal/ecclient"
-	"github.com/elastic/apm-server/functionaltests/internal/esclient"
-)
-
-// ingestV7Step performs ingestion to the APM Server deployed on ECH.
-// After ingestion, it checks if the document counts difference between
-// current and previous is expected.
-//
-// The output of this step is the indices document counts after ingestion.
-//
-// NOTE: Only works for versions 7.x.
-type ingestV7Step struct{}
-
-func (i ingestV7Step) Step(t *testing.T, ctx context.Context, e *testStepEnv) {
-	if e.currentVersion().Major >= 8 {
-		t.Fatal("ingest v7 step should only be used for versions < 8.0")
-	}
-
-	var beforeIngestDSDocCount esclient.DataStreamsDocCount
-	var beforeIngestIdxDocCount esclient.IndicesDocCount
-	if e.integrations {
-		beforeIngestDSDocCount = getDocCountPerDSV7(t, ctx, e.esc, e.dsNamespace)
-	} else {
-		beforeIngestIdxDocCount = getDocCountPerIndexV7(t, ctx, e.esc)
-	}
-
-	t.Logf("------ ingest in %s ------", e.currentVersion())
-	err := e.gen.RunBlockingWait(ctx, e.currentVersion(), e.integrations)
-	require.NoError(t, err)
-
-	t.Logf("------ ingest check in %s ------", e.currentVersion())
-	t.Log("check number of documents increased after ingestion")
-	if e.integrations {
-		// Managed, check data streams.
-		afterIngestDSDocCount := getDocCountPerDSV7(t, ctx, e.esc, e.dsNamespace)
-		asserts.DocExistFor(t, afterIngestDSDocCount, expectedDataStreams(e.dsNamespace))
-		asserts.DocCountIncreased(t, afterIngestDSDocCount, beforeIngestDSDocCount)
-	} else {
-		// Standalone, check indices.
-		afterIngestIdxDocCount := getDocCountPerIndexV7(t, ctx, e.esc)
-		asserts.DocExistFor(t, afterIngestIdxDocCount, expectedIndices())
-		asserts.DocCountIncreased(t, afterIngestIdxDocCount, beforeIngestIdxDocCount)
-	}
-}
-
-// upgradeV7Step upgrades the ECH deployment from its current version to
-// the new version. It also adds the new version into testStepEnv. After
-// upgrade, it checks that the document counts did not change across upgrade.
-//
-// The output of this step is the indices document counts if upgrading to 7.x,
-// or data streams document counts if upgrading to >= 8.0.
-//
-// NOTE: Only works from versions 7.x.
-type upgradeV7Step struct {
-	NewVersion ecclient.StackVersion
-}
-
-func (u upgradeV7Step) Step(t *testing.T, ctx context.Context, e *testStepEnv) {
-	if e.currentVersion().Major >= 8 {
-		t.Fatal("upgrade v7 step should only be used from versions < 8.0")
-	}
-
-	var beforeUpgradeDSDocCount esclient.DataStreamsDocCount
-	var beforeUpgradeIdxDocCount esclient.IndicesDocCount
-	if e.integrations {
-		beforeUpgradeDSDocCount = getDocCountPerDSV7(t, ctx, e.esc, e.dsNamespace)
-	} else {
-		beforeUpgradeIdxDocCount = getDocCountPerIndexV7(t, ctx, e.esc)
-	}
-
-	t.Logf("------ upgrade %s to %s ------", e.currentVersion(), u.NewVersion)
-	upgradeCluster(t, ctx, e.tf, e.target, u.NewVersion, e.integrations)
-	// Update the environment version to the new one.
-	e.versions = append(e.versions, u.NewVersion)
-
-	t.Logf("------ upgrade check in %s ------", e.currentVersion())
-	t.Log("check number of documents stayed the same across upgrade")
-	// We assert that no changes happened in the number of documents after upgrade
-	// to ensure the state didn't change.
-	// We don't expect any change here unless something broke during the upgrade.
-	if e.integrations {
-		// Managed, check data streams.
-		afterUpgradeDSDocCount := getDocCountPerDSV7(t, ctx, e.esc, e.dsNamespace)
-		asserts.DocExistFor(t, afterUpgradeDSDocCount, expectedDataStreams(e.dsNamespace))
-		asserts.DocCountStayedTheSame(t, afterUpgradeDSDocCount, beforeUpgradeDSDocCount)
-	} else {
-		// Standalone, check indices.
-		afterUpgradeIdxDocCount := getDocCountPerIndexV7(t, ctx, e.esc)
-		asserts.DocExistFor(t, afterUpgradeIdxDocCount, expectedIndices())
-		asserts.DocCountStayedTheSame(t, afterUpgradeIdxDocCount, beforeUpgradeIdxDocCount)
-	}
-}
-
-// migrateManagedStep migrates the ECH APM deployment from standalone mode to
-// managed mode, which involves enabling the integrations server via Kibana.
-// It also checks that the document counts did not change across the migration.
-//
-// The output of this step is the indices document counts if version < 8.0,
-// or data streams document counts if version >= 8.0.
-type migrateManagedStep struct{}
-
-func (m migrateManagedStep) Step(t *testing.T, ctx context.Context, e *testStepEnv) {
-	if e.integrations {
-		t.Fatal("migrate managed step should only be used on standalone")
-	}
-
-	var beforeMigrateDSDocCount esclient.DataStreamsDocCount
-	var beforeMigrateIdxDocCount esclient.IndicesDocCount
-	if e.currentVersion().Major >= 8 {
-		beforeMigrateDSDocCount = getDocCountPerDSV7(t, ctx, e.esc, e.dsNamespace)
-	} else {
-		beforeMigrateIdxDocCount = getDocCountPerIndexV7(t, ctx, e.esc)
-	}
-
-	t.Logf("------ migrate to managed for %s ------", e.currentVersion())
-	t.Log("enable integrations server")
-	err := e.kbc.EnableIntegrationsServer(ctx)
-	require.NoError(t, err)
-	e.integrations = true
-
-	// APM Server needs some time to start serving requests again, and we don't have any
-	// visibility on when this completes.
-	// NOTE: This value comes from empirical observations.
-	time.Sleep(80 * time.Second)
-
-	t.Log("check number of documents stayed the same across migration to managed")
-	// We assert that no changes happened in the number of documents after migration
-	// to ensure the state didn't change.
-	// We don't expect any change here unless something broke during the migration.
-	if e.currentVersion().Major >= 8 {
-		afterMigrateDSDocCount := getDocCountPerDS(t, ctx, e.esc)
-		asserts.DocExistFor(t, afterMigrateDSDocCount, expectedDataStreams(e.dsNamespace))
-		asserts.DocCountStayedTheSame(t, afterMigrateDSDocCount, beforeMigrateDSDocCount)
-	} else {
-		afterMigrateIdxDocCount := getDocCountPerIndexV7(t, ctx, e.esc)
-		asserts.DocExistFor(t, afterMigrateIdxDocCount, expectedIndices())
-		asserts.DocCountStayedTheSame(t, afterMigrateIdxDocCount, beforeMigrateIdxDocCount)
-	}
-}
-
-// resolveDeprecationsStep resolves critical migration deprecation warnings from Elasticsearch regarding
-// indices created in 7.x not being compatible with 9.x.
-//
-// The output of this step is the previous test step result.
-type resolveDeprecationsStep struct{}
-
-func (r resolveDeprecationsStep) Step(t *testing.T, ctx context.Context, e *testStepEnv) {
-	t.Logf("------ resolve migration deprecations in %s ------", e.currentVersion())
-	err := e.kbc.ResolveMigrationDeprecations(ctx)
-	require.NoError(t, err)
-}
-
-func expectedIndices() []string {
-	return []string{
-		"apm-*-error-*",
-		"apm-*-span-*",
-		"apm-*-transaction-*",
-		"apm-*-metric-*",
-	}
-}
-
-// getDocCountPerDSV7 retrieves document count per data stream for versions < 8.0.
-func getDocCountPerDSV7(t *testing.T, ctx context.Context, esc *esclient.Client, namespace string) esclient.DataStreamsDocCount {
-	t.Helper()
-	count, err := esc.APMDSDocCountV7(ctx, namespace)
-	require.NoError(t, err)
-	return count
-}
-
-// getDocCountPerIndexV7 retrieves document count per index for versions < 8.0.
-func getDocCountPerIndexV7(t *testing.T, ctx context.Context, esc *esclient.Client) esclient.IndicesDocCount {
-	t.Helper()
-	count, err := esc.APMIdxDocCountV7(ctx)
-	require.NoError(t, err)
-	return count
-}
